+++ http/src/transport/boxed_transport.rs
@@ -6,9 +6,8 @@ use std::{
     io::Result,
     net::SocketAddr,
     ops::Deref,
-    pin::Pin,
-    task::{Context, Poll},
 };
+use trillium_macros::{AsyncRead, AsyncWrite};
 
 #[allow(missing_docs)]
 #[doc(hidden)]
@@ -41,13 +40,12 @@ throughout an application.
 replaced by a type alias exported from a
 `trillium_server_common::Server` crate.
 */
+#[derive(Debug, AsyncRead, AsyncWrite)]
 pub struct BoxedTransport(Box<dyn AnyTransport>);
 
-impl Debug for BoxedTransport {
+impl Debug for Box<dyn AnyTransport> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("BoxedTransport")
-            .field("inner", &"Box<dyn Transport>")
-            .finish()
+        f.write_str("Box<dyn Any + Transport>")
     }
 }
 
@@ -94,34 +92,6 @@ impl Deref for BoxedTransport {
 
     fn deref(&self) -> &Self::Target {
         self.0.as_transport()
-    }
-}
-
-impl AsyncRead for BoxedTransport {
-    fn poll_read(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut [u8],
-    ) -> Poll<Result<usize>> {
-        Pin::new(&mut self.0).poll_read(cx, buf)
-    }
-}
-
-impl AsyncWrite for BoxedTransport {
-    fn poll_write(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &[u8],
-    ) -> Poll<Result<usize>> {
-        Pin::new(&mut self.0).poll_write(cx, buf)
-    }
-
-    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<()>> {
-        Pin::new(&mut self.0).poll_flush(cx)
-    }
-
-    fn poll_close(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<()>> {
-        Pin::new(&mut self.0).poll_close(cx)
     }
 }
