+++ smol/Cargo.toml
@@ -18,6 +18,7 @@ futures-lite = "1.12.0"
 log = "0.4.17"
 trillium = { path = "../trillium", version = "^0.2.0" }
 trillium-http = { path = "../http", version = "^0.2.0" }
+trillium-macros = { version = "0.0.3", path = "../macros" }
 trillium-server-common = { path = "../server-common", version = "^0.3.0" }
 
 [target.'cfg(unix)'.dependencies]
