+++ tokio/Cargo.toml
@@ -16,6 +16,7 @@ log = "0.4.17"
 tokio-stream = { version = "0.1.12", features = ["net"] }
 trillium = { path = "../trillium", version = "^0.2.0" }
 trillium-http = { path = "../http", version = "^0.2.0" }
+trillium-macros = { version = "0.0.3", path = "../macros" }
 trillium-server-common = { path = "../server-common", version = "^0.3.0" }
 
 [dependencies.tokio]
