+++ tokio/src/transport.rs
@@ -1,15 +1,11 @@
 use async_compat::Compat;
-use std::{
-    io::Result,
-    net::SocketAddr,
-    pin::Pin,
-    task::{Context, Poll},
-};
+use std::{io::Result, net::SocketAddr};
 use tokio::net::{TcpStream, ToSocketAddrs};
+use trillium_macros::{AsyncRead, AsyncWrite};
 use trillium_server_common::{AsyncRead, AsyncWrite, Transport};
 
 /// A transport newtype for tokio
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, AsyncRead, AsyncWrite)]
 pub struct TokioTransport<T>(pub(crate) T);
 
 impl<T> TokioTransport<T> {
@@ -31,40 +27,6 @@ impl TokioTransport<Compat<TcpStream>> {
 impl<T> From<T> for TokioTransport<T> {
     fn from(value: T) -> Self {
         Self(value)
-    }
-}
-
-impl<T> AsyncRead for TokioTransport<T>
-where
-    T: Unpin + AsyncRead,
-{
-    fn poll_read(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut [u8],
-    ) -> Poll<Result<usize>> {
-        Pin::new(&mut self.0).poll_read(cx, buf)
-    }
-}
-
-impl<T> AsyncWrite for TokioTransport<T>
-where
-    T: Unpin + AsyncWrite,
-{
-    fn poll_write(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &[u8],
-    ) -> Poll<Result<usize>> {
-        Pin::new(&mut self.0).poll_write(cx, buf)
-    }
-
-    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<()>> {
-        Pin::new(&mut self.0).poll_flush(cx)
-    }
-
-    fn poll_close(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<()>> {
-        Pin::new(&mut self.0).poll_close(cx)
     }
 }
